backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pandas==2.2.2
PyPDF2==3.0.1
docx2txt==0.8
python-multipart==0.0.9
faiss-cpu==1.8.0
sentence-transformers==3.1.1
requests==2.32.3
numpy==1.26.4